Rationalise header file inclusion
Header files themselves should only include those header files
that are needed to use the header file itself. Other includes for
the module should be in the c source file.

C source files should include their own header file.

There were many header files (both system and application) that
were included unnecessarily, and these have now been removed.

+++ lib/notify.c
@@ -27,12 +27,10 @@
 #endif
 #include <unistd.h>
 #include <stdlib.h>
-#include <syslog.h>
 #include <fcntl.h>
 #include <errno.h>
 #include <signal.h>
 #include <grp.h>
-#include <sys/types.h>
 #include <string.h>
 #include <sys/stat.h>
 #include <pwd.h>
@@ -41,7 +39,6 @@
 #include "signals.h"
 #include "logger.h"
 #include "utils.h"
-#include "vector.h"
 #include "parser.h"
 
 /* Default user/group for script execution */
@@ -259,7 +256,7 @@ find_path(notify_script_t *script, bool full_string)
 		path = getenv ("PATH");
 
 		if (!path) {
-			int cs_path_len;
+			size_t cs_path_len;
 			path = MALLOC(cs_path_len = confstr(_CS_PATH, NULL, 0));
 			confstr(_CS_PATH, path, cs_path_len);
 			path_is_malloced = true;
@@ -287,7 +284,7 @@ find_path(notify_script_t *script, bool full_string)
 
 		/* Get our supplementary groups */
 		sgid_num = getgroups(0, NULL);
-		sgid_list = MALLOC((sgid_num + 1) * sizeof(gid_t));
+		sgid_list = MALLOC((size_t)(sgid_num + 1) * sizeof(gid_t));
 		sgid_num = getgroups(sgid_num, sgid_list);
 		sgid_list[sgid_num++] = 0;
 
@@ -324,7 +321,7 @@ find_path(notify_script_t *script, bool full_string)
 		}
 
 		/* Use the current path entry, plus a '/' if nonempty, plus the file to execute. */
-		char *pend = mempcpy (buffer, p, subp - p);
+		char *pend = mempcpy (buffer, p, (size_t)(subp - p));
 		*pend = '/';
 		memcpy (pend + (p < subp), file, file_len + 1);
 
@@ -404,7 +401,7 @@ exit:
 
 		/* restore supplementary groups */
 		if (sgid_list) {
-			if (setgroups(sgid_num, sgid_list))
+			if (setgroups((size_t)sgid_num, sgid_list))
 				log_message(LOG_INFO, "Unable to restore supplementary groups after script search (%m)");
 			FREE(sgid_list);
 		}
